configs: msm8976: enable fstack-protector-strong
Upgrade protection against stack corruption from
CONFIG_CC_STACKPROTECTOR_REGULAR TO CONFIG_CC_STACKPROTECT_STRONG.
(most) Functions with the possibility of stack corruption are
protected with a canary.

With this change, the marlin kernel grows from 15633096 to 15837372
an increase of 1.3%. Performance loss is considered minimal. Security
against stack overflow is greatly improved. [1]

This is the improved version of CONFIG_CC_STACKPROTECTOR=y which is
enabled in shamu's kernel.

[1] https://lwn.net/Articles/584225

Bug: 28837708
Change-Id: I79229f0a492be9f60f09faf265871c1cbe6a3e42